+++ src/main/java/gabriel/betbot/dtos/tradefeed/Result.java
@@ -5,6 +5,7 @@ import java.util.List;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+import java.math.BigInteger;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
@@ -14,11 +15,11 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 public class Result {
 
     @JsonProperty("Since")
-    public final Integer since;
+    public final Long since;
     @JsonProperty("Sports")
     public final List<Sport> sports;
 
-    public Result(final Integer since, final List<Sport> sports) {
+    public Result(final Long since, final List<Sport> sports) {
         this.since = since;
         this.sports = sports;
     }
@@ -27,5 +28,7 @@ public class Result {
         this.since = null;
         this.sports = null;
     }
+      
+  
     
 }
