+++ src/main/java/gabriel/betbot/dtos/tradefeed/TradeFeedDto.java
@@ -4,6 +4,7 @@ package gabriel.betbot.dtos.tradefeed;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+import java.math.BigInteger;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
@@ -14,13 +15,13 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 public class TradeFeedDto {
 
     @JsonProperty("Code")
-    public final Integer code;
+    public final Long code;
     @JsonProperty("Message")
     public final String message;
     @JsonProperty("Result")
     public final Result result;
 
-    public TradeFeedDto(final Integer code, final String message, final Result result) {
+    public TradeFeedDto(final Long code, final String message, final Result result) {
         this.code = code;
         this.message = message;
         this.result = result;
